branch old_raknet in trunk gemergt. jetzt wird Raknet 2.x benoetigt

+++ src/core/charconv.h
@@ -1,9 +1,18 @@
 #ifndef CHARCONV_H
 #define CHARCONV_H
-#include "raknet/RakNetTypes.h"
+/*
+#include "RakNetTypes.h"
 #include "raknet/BitStream.h"
 #include "raknet/GetTime.h"
 #include "raknet/MessageIdentifiers.h"
+*/
+#include "raknet/PacketEnumerations.h"
+#include "raknet/RakNetworkFactory.h"
+#include "raknet/RakPeerInterface.h"
+#include "raknet/NetworkTypes.h"
+#include "raknet/BitStream.h"
+#include "raknet/GetTime.h"
+
 #include <algorithm>
 #include <string>
 #include <cstring>
@@ -16,26 +25,26 @@
 class CharConv
 {
 	public:
-	
+
 	/**
 	 * \fn CharConv()
 	 * \brief Konstruktor
 	 */
 	CharConv();
-	
+
 	/**
 	 * \fn CharConv(int dummy)
 	 * \brief Erzeugt ein Objekt mit leerem Bitstream (Standard ist, dass eine Kennung und ein Zeitstempel geschrieben wird)
 	 * \param dummy Dummy Parameter um eine unterschiedliche Signatur zu erzeugen
 	 */
 	CharConv(int dummy);
-	
+
 	/**
 	 * \fn CharConv(std::iostream* stream)
 	 * \param stream Datenstrom, in den Daten geschrieben werden
 	 */
 	CharConv(std::iostream* stream);
-	
+
 	/**
 	 * \fn CharConv(unsigned char* data, unsigned int size)
 	 * \brief erstellt ein neues Objekt auf der Basis der angegebenen Daten
@@ -43,15 +52,15 @@ class CharConv
 	 * \param size Laenge der Zeichenkette data
 	 */
 	CharConv(unsigned char* data, unsigned int size);
-	
+
 	/**
 	 * \fn CharConv(Packet* packet)
 	 * \brief erstellt ein neues Objekt aus einem ueber das Netzwerk erhaltenen Datenpaket
 	 * \param packet Datenpaket
 	 */
 	CharConv(Packet* packet);
-	
-	
+
+
 	/**
 	 * \fn void backToStart()
 	 * \brief setzt den internen Lesezeiger zurueck an den Anfang
@@ -60,7 +69,7 @@ class CharConv
 	{
 		m_bitstream.SetWriteOffset(0);
 	}
-	
+
 	/**
 	 * \fn void reset()
 	 * \brief Loescht die zugrunde liegenden Daten
@@ -69,7 +78,7 @@ class CharConv
 	{
 		m_bitstream.Reset();
 	}
-	
+
 	/**
 	 * \fn void toBuffer(const char* data, unsigned int size)
 	 * \brief Schreibt Daten in den Puffer
@@ -77,8 +86,8 @@ class CharConv
 	 * \param size Laenge der zu schreibenden Daten
 	 */
 	void toBuffer(const char* data, unsigned int size);
-	
-	
+
+
 	/**
 	 * \fn void fromBuffer(char* data, unsigned int size)
 	 * \brief Liest Daten als String aus dem Puffer
@@ -86,7 +95,7 @@ class CharConv
 	 * \param size Anzahl der Zeichen die gelesen werden
 	 */
 	void fromBuffer(char* data, unsigned int size);
-	
+
 	/**
 	 * \fn void toBuffer(std::string s, unsigned int size)
 	 * \brief Schreibt aus dem String die ersten size Zeichen in den Puffer
@@ -94,8 +103,8 @@ class CharConv
 	 * \param size Anzahl Zeichen
 	 */
 	void toBuffer(std::string s, unsigned int size);
-	
-	
+
+
 	/**
 	 * \fn void fromBuffer(std::string& s, unsigned int size)
 	 * \brief Liest aus dem String size Zeichen und schreibt sie in den String
@@ -103,8 +112,8 @@ class CharConv
 	 * \param size Anzahl Zeichen
 	 */
 	void fromBuffer(std::string& s, unsigned int size);
-	
-	
+
+
 	/**
 	 * \fn void toBuffer(T data)
 	 * \brief Schreibt Daten in den Puffer
@@ -122,7 +131,7 @@ class CharConv
 			(*m_stream) << data << " ";
 		}
 	}
-	
+
 	/**
 	 * \fn void fromBuffer(T &data)
 	 * \brief Liest Daten aus dem Puffer
@@ -140,28 +149,28 @@ class CharConv
 			(*m_stream) >> data;
 		}
 	}
-	
-	
-	
-	
+
+
+
+
 	/**
 	 * \fn void toBuffer(std::string s)
 	 * \brief Schreibt einen String in den Buffer, indem zuerst die Laenge und dann die Daten geschrieben werden
 	 * \param s String
 	 */
 	void toBuffer(std::string s);
-	
-	
+
+
 	/**
 	 * \fn void fromBuffer(std::string s)
 	 * \brief liest einen String aus dem Puffer
 	 * \param s String
 	 */
 	void fromBuffer(std::string& s);
-	
-	
-	
-	
+
+
+
+
 	/**
 	 * \fn RakNet::BitStream* getBitStream()
 	 * \brief Gibt den Bitstream der die Daten enthaelt aus
@@ -170,7 +179,7 @@ class CharConv
 	{
 		return &m_bitstream;
 	}
-	
+
 	/**
 	 * \fn std::iostream* getStream()
 	 * \brief  Gibt den Stream aus
@@ -179,7 +188,7 @@ class CharConv
 	{
 		return m_stream;
 	}
-	
+
 	/**
 	 * \fn int readBits()
 	 * \brief Gibt die Anzahl der bisher gelesenen Bits aus
@@ -188,7 +197,7 @@ class CharConv
 	{
 		return m_bitstream.GetReadOffset();
 	}
-	
+
 	/**
 	 * \fn int writeBits()
 	 * \brief Gibt die Anzahl der bisher geschriebenen Bits aus
@@ -197,7 +206,7 @@ class CharConv
 	{
 		return m_bitstream.GetNumberOfBitsUsed();
 	}
-	
+
 	/**
 	 * \fn float getDelay()
 	 * \brief Gibt die Verzoegerung des Paketes in Millisekunden an
@@ -206,16 +215,15 @@ class CharConv
 	{
 		return std::max(0.0f,static_cast<float>(RakNet::GetTime())-static_cast<float>(m_timestamp));
 	}
-	
+
 	/**
-	 * \fn unsigned long getTimestamp()
 	 * \brief Gibt den Zeitstempel des Paketes aus
 	 */
-	unsigned long getTimestamp()
+	unsigned int getTimestamp()
 	{
 		return m_timestamp;
 	}
-	
+
 	/**
 	 * \fn void printNewline()
 	 * \brief Schreibt Newline (Debugging)
@@ -225,7 +233,7 @@ class CharConv
 		if (m_stream !=0)
 			(*m_stream) << "\n";
 	}
-	
+
 	/**
 	 * \fn  int getId()
 	 * \brief Gibt die ID aus
@@ -234,7 +242,7 @@ class CharConv
 	{
 		return m_id;
 	}
-	
+
 	/**
 	 * \fn void setId(int id)
 	 * \brief Setzt die ID
@@ -244,7 +252,7 @@ class CharConv
 	{
 		m_id = id;
 	}
-	
+
 	/**
 	 * \fn  int getVersion()
 	 * \brief Gibt die Version aus
@@ -253,7 +261,7 @@ class CharConv
 	{
 		return m_version;
 	}
-	
+
 	/**
 	 * \fn void setVersion(int version)
 	 * \brief Setzt die Version
@@ -263,39 +271,39 @@ class CharConv
 	{
 		m_version = version;
 	}
-	
-	
+
+
 	private:
 		/**
 		 * \var RakNet::BitStream m_bitstream
 		 * \brief enthaelt die Daten als Zeichenkette
 		 */
 		RakNet::BitStream m_bitstream;
-		
+
 		/**
 		 * \var unsigned long m_timestamp
 		 * \brief Zeit zur das Paket erstellt bzw erhalten wurde
 		 */
-		unsigned long m_timestamp;
-		
+		unsigned int m_timestamp;
+
 		/**
 		 * \var std::iostream* m_stream
 		 * \brief Ein/Ausgabestrom - wenn nicht auf 0 gesetzt wird dieser Strom verwendet
 		 */
 		std::iostream* m_stream;
-		
+
 		/**
 		 * \var int m_id
 		 * \brief ID
 		 */
 		int m_id;
-		
+
 		/**
 		 * \var int m_version
 		 * \brief Versionsnummer der Daten, die gerade geschrieben oder gelesen werden
 		 */
 		int m_version;
-	
+
 };
 
 /**
@@ -305,7 +313,7 @@ class CharConv
  */
 template <>
 void CharConv::toBuffer<char>(char c);
-	
+
 /**
  * \fn void fromBuffer(char &c)
  * \brief liest einen char aus dem Puffer
